test: Change type of iterations parameter to u32

+++ crates/testing/test-programs/fibo/workflow/src/lib.rs
@@ -12,7 +12,7 @@ pub fn black_box<T>(dummy: T) -> T {
 }
 
 impl Guest for Component {
-    fn fibow(n: u8, iterations: u16) -> u64 {
+    fn fibow(n: u8, iterations: u32) -> u64 {
         let mut last = 0;
         for _ in 0..iterations {
             last = fibo(black_box(n));
@@ -20,7 +20,7 @@ impl Guest for Component {
         last
     }
 
-    fn fiboa(n: u8, iterations: u16) -> u64 {
+    fn fiboa(n: u8, iterations: u32) -> u64 {
         let mut last = 0;
         for _ in 0..iterations {
             last = crate::bindings::testing::fibo::fibo::fibo(n);
